Added location handling of permission
