+++ src/App.js
@@ -53,27 +53,6 @@ function isLandingFirstTime() {
   return !localStorage.getItem("firstTime");
 }
 
-function getUserCamera(handle, onSuccess, onError) {
-  const constraints = (window.constraints = {
-    audio: false,
-    video: true
-  });
-
-  if (navigator.mediaDevices) {
-    navigator.mediaDevices
-      .getUserMedia(constraints)
-      .then(stream => {
-        onSuccess();
-      })
-      .catch(err => {
-        onError(err);
-      });
-  } else {
-    console.log("Browser does not support media devices API");
-    alert("Your browser does not support the use of media");
-  }
-}
-
 const App = () => {
   const history = useHistory();
   const permissions = useQuery(PERMISSIONS);
