+++ src/utils/UserLocation.js
@@ -16,13 +16,17 @@ const gpsMutationHook = () => {
   return useMutation(UPDATE_GPS);
 };
 
-const setGpsCache = hook => {
+const setGpsCache = (hook, onSuccess, onError) => {
   navigator.geolocation.getCurrentPosition(
     position => {
       const { latitude, longitude } = position.coords;
       hook({ variables: { latitude, longitude } });
+      if (onSuccess) onSuccess(position);
     },
-    err => console.error(err),
+    err => {
+      console.error(err);
+      if (onError) onError(err);
+    },
     {
       timeout: 5000
     }
